fix: Update Session Management to use Long IDs and fix entity compatibility
Changes:
- Updated Session entity ID from Integer to Long for consistency with User/Conference
- Added deleted boolean field for soft delete support
- Updated SessionRepository with all required query methods
- Fixed SessionService to use Long IDs and LocalDateTime
- Updated all DTOs (CreateSessionRequest, SessionResponse, etc.) to use Long
- Updated Room entity ID to Long
- Fixed SessionRepository with availability checking queries
- Removed references to non-existent entity methods (setDeletedAt, getVersion)

In Progress:
- DTO method signatures still need complete Integer->Long conversion
- AdminController needs updates for DTO compatibility
- Response DTOs (RoomAvailabilityResponse, TimeConflictResponse) need fixes

Disabled temporarily:
- RegistrationService, RoomService, ScheduleService (incomplete dependencies)

+++ src/main/java/com/team5/demo/dto/ConferenceDTO.java
@@ -1,7 +1,7 @@
 package com.team5.demo.dto;
 
 public class ConferenceDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
 
@@ -9,7 +9,7 @@ public class ConferenceDTO {
     public ConferenceDTO() {
     }
 
-    public ConferenceDTO(Integer id, String title, String description) {
+    public ConferenceDTO(Long id, String title, String description) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -20,7 +20,7 @@ public class ConferenceDTO {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
