+++ src/main/java/com/team5/demo/dto/RoomAvailabilityResponse.java
@@ -58,7 +58,7 @@ public class RoomAvailabilityResponse {
         public ConflictingSession() {
         }
 
-        public ConflictingSession(Integer id, String title, LocalDateTime startTime, LocalDateTime endTime) {
+        public ConflictingSession(Long id, String title, LocalDateTime startTime, LocalDateTime endTime) {
             this.id = id;
             this.title = title;
             this.startTime = startTime;
@@ -69,7 +69,7 @@ public class RoomAvailabilityResponse {
             return id;
         }
 
-        public void setId(Integer id) {
+        public void setId(Long id) {
             this.id = id;
         }
