+++ src/main/java/com/team5/demo/dto/UserDTO.java
@@ -1,7 +1,7 @@
 package com.team5.demo.dto;
 
 public class UserDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
@@ -9,7 +9,7 @@ public class UserDTO {
     public UserDTO() {
     }
 
-    public UserDTO(Integer id, String name, String email) {
+    public UserDTO(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -20,7 +20,7 @@ public class UserDTO {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
